+++ tests/testthat/test-relist-nlist.R
@@ -1,0 +1,8 @@
+context("relist")
+
+test_that("relist_nlist", {
+  expect_identical(relist_nlist(structure(numeric(0), .Names = character(0)), 
+                                nlist()), nlist())
+  expect_identical(relist_nlist(c(`a[2]` = 5), nlist(a = c(1, 2, 3))),
+                   nlist(a = c(NA, 5, NA)))
+})