+++ NEWS.md
@@ -1,6 +1,6 @@
 - Added `fill_na()` for natomic, nlist and nlists objects.
 - Added `as.nlist.numeric()` to convert flattened nlist objects back into original objects.
-- Added `unlist.nlist()` to flatten nlist objects into term named numeric vectors.
+- Added `unlist_nlist()` (also `unlist.nlist()`) and `relist_nlist()` to flatten and pop nlist objects.
 
 # nlist 0.0.1